Moved Saml code to S.IM.Tokens. Removed a dependency.
