+++ src/Microsoft.IdentityModel.Protocol.Extensions/project.json
@@ -7,12 +7,6 @@
     },
     "code": [ "**\\*.cs", "..\\common\\*.cs" ],
     "exclude": [
-        "SecurityKeyResolver.cs",
-        "SecurityTokenHandlerCollectionExtensions.cs",
-        "Saml2SecurityTokenHandler.cs",
-        "SamlSecurityTokenHandler.cs",
-        "Saml2SecurityToken.cs",
-        "SamlSecurityToken.cs",
         "WsFederationMessage.cs",
         "Configuration\\WsFederationConfiguration.cs",
         "Configuration\\WsFederationConfigurationRetriever.cs"
