+++ src/System.IdentityModel.Tokens/Saml/Saml2SecurityTokenHandler.cs
@@ -6,9 +6,9 @@
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
-// 
+//
 // http://www.apache.org/licenses/LICENSE-2.0
-// 
+//
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -20,40 +20,58 @@ using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.Globalization;
+using System.IdentityModel.Selectors;
 using System.IdentityModel.Tokens;
 using System.IO;
 using System.Security.Claims;
 using System.Text;
 using System.Xml;
-using SamlHandler = System.IdentityModel.Tokens.SamlSecurityTokenHandler;
+using Saml2Handler = System.IdentityModel.Tokens.Saml2SecurityTokenHandler;
 
 namespace Microsoft.IdentityModel.Tokens
 {
     /// <summary>
-    /// A derived <see cref="System.IdentityModel.Tokens.Saml2SecurityTokenHandler"/> that implements ISecurityTokenValidator, 
+    /// A derived <see cref="System.IdentityModel.Tokens.Saml2SecurityTokenHandler"/> that implements ISecurityTokenValidator,
     /// which supports validating tokens passed as strings using <see cref="TokenValidationParameters"/>.
     /// </summary>
-    ///     
-    public class SamlSecurityTokenHandler : SecurityTokenHandler, ISecurityTokenValidator
+    ///
+    public class Saml2SecurityTokenHandler : SecurityTokenHandler, ISecurityTokenValidator
     {
-        internal const string SamlTokenProfile11 = "urn:oasis:names:tc:SAML:1.0:assertion";
-        internal const string OasisWssSamlTokenProfile11 = "http://docs.oasis-open.org/wss/oasis-wss-saml-token-profile-1.1#SAMLV1.1";
+        internal const string Saml2TokenProfile11 = "urn:oasis:names:tc:SAML:2.0:assertion";
+        internal const string OasisWssSaml2TokenProfile11 = "http://docs.oasis-open.org/wss/oasis-wss-saml-token-profile-1.1#SAMLV2.0";
 
         private Int32 _maximumTokenSizeInBytes = TokenValidationParameters.DefaultMaximumTokenSizeInBytes;
-        private static string[] _tokenTypeIdentifiers = new string[] { SamlTokenProfile11, OasisWssSamlTokenProfile11 };
+        private static string[] _tokenTypeIdentifiers = new string[] { Saml2TokenProfile11, OasisWssSaml2TokenProfile11 };
 
         // never set any properties on the handler.
-        private static SMSamlHandlerPrivate _smSamlHandlerPrivateNeverSetAnyProperties = new SMSamlHandlerPrivate();
+        private static SMSaml2HandlerPrivate _smSaml2HandlerPrivateNeverSetAnyProperties = new SMSaml2HandlerPrivate();
 
         /// <summary>
-        /// Initializes an instance of <see cref="SamlSecurityTokenHandler"/>
+        /// Initializes a new instance of <see cref="Saml2SecurityTokenHandler"/>.
         /// </summary>
-        public SamlSecurityTokenHandler()
-        {        
+        public Saml2SecurityTokenHandler()
+        {
         }
 
         /// <summary>
-        /// Gets a value indicating whether this handler supports validation of tokens 
+        /// Gets the token type identifier(s) supported by this handler.
+        /// </summary>
+        /// <returns>A collection of strings that identify the tokens this instance can handle.</returns>
+        public override string[] GetTokenTypeIdentifiers()
+        {
+            return _tokenTypeIdentifiers;
+        }
+
+        /// <summary>
+        /// Gets the token type supported by this handler.
+        /// </summary>
+        public override Type TokenType
+        {
+            get { return typeof(Saml2SecurityToken); }
+        }
+
+        /// <summary>
+        /// Gets a value indicating whether this handler supports validation of tokens
         /// handled by this instance.
         /// </summary>v
         /// <returns>'True' if the instance is capable of SecurityToken
@@ -64,20 +82,21 @@ namespace Microsoft.IdentityModel.Tokens
         }
 
         /// <summary>
-        /// Gets a value indicating whether the class provides serialization functionality to serialize securityToken handled
+        /// Gets a value indicating whether the class provides serialization functionality to serialize the token handled
         /// by this instance.
         /// </summary>
-        /// <returns>true if the WriteToken method can serialize this securityToken.</returns>
+        /// <returns>true if the WriteToken method can serialize this token.</returns>
         public override bool CanWriteToken
         {
             get { return true; }
         }
 
         /// <summary>
-        /// Reads the string as XML and looks for the an element <see cref="SamlConstants.Assertion"/> with namespace <see cref="SamlConstants.Saml11Namespace"/>. 
+        /// Reads the string as XML and looks for the an element <see cref="SamlConstants.Assertion"/> or  <see cref="SamlConstants.EncryptedAssertion"/> with namespace <see cref="SamlConstants.Saml2Namespace"/>.
         /// </summary>
         /// <param name="securityToken">The securitytoken.</param>
-        /// <returns><see cref="XmlDictionaryReader.IsStartElement"/> (<see cref="SamlConstants.Assertion"/>, <see cref="SamlConstants.Saml11Namespace"/>).</returns>
+        /// <returns><see cref="XmlDictionaryReader.IsStartElement"/> (<see cref="SamlConstants.Assertion"/>, <see cref="SamlConstants.Saml2Namespace"/>)
+        /// OR <see cref="XmlDictionaryReader.IsStartElement"/> (<see cref="SamlConstants.EncryptedAssertion"/>, <see cref="SamlConstants.Saml2Namespace"/>).</returns>
         public override bool CanReadToken(string securityToken)
         {
             if (string.IsNullOrWhiteSpace(securityToken) || securityToken.Length > MaximumTokenSizeInBytes)
@@ -97,23 +116,47 @@ namespace Microsoft.IdentityModel.Tokens
                     {
                         return false;
                     }
-                    return reader.IsStartElement(SamlConstants.Assertion, SamlConstants.Saml11Namespace);
+
+                    return reader.IsStartElement(SamlConstants.Assertion, SamlConstants.Saml2Namespace)
+                        || reader.IsStartElement(SamlConstants.EncryptedAssertion, SamlConstants.Saml2Namespace);
                 }
             }
         }
 
         /// <summary>
-        /// Creates claims from a Saml securityToken.
+        /// Creates the security token reference when the token is not attached to the message.
         /// </summary>
-        /// <param name="samlToken">A <see cref="SamlSecurityToken"/> that will be used to create the claims.</param>
-        /// <param name="issuer">the issuer value for each <see cref="Claim"/> in the <see cref="ClaimsIdentity"/>.</param>/// 
-        /// <param name="validationParameters"> contains parameters for validating the securityToken.</param>
-        /// <returns>A <see cref="ClaimsIdentity"/> containing the claims from the <see cref="SamlSecurityToken"/>.</returns>
-        protected virtual ClaimsIdentity CreateClaimsIdentity(SamlSecurityToken samlToken, string issuer, TokenValidationParameters validationParameters)
+        /// <param name="token">The saml token.</param>
+        /// <param name="attached">Boolean that indicates if a attached or unattached
+        /// reference needs to be created.</param>
+        /// <returns>A <see cref="Saml2AssertionKeyIdentifierClause"/>.</returns>
+        public override SecurityKeyIdentifierClause CreateSecurityTokenReference(SecurityToken token, bool attached)
+        {
+            if (null == token)
+            {
+                throw new ArgumentNullException("token");
+            }
+
+            return token.CreateKeyIdentifierClause<Saml2AssertionKeyIdentifierClause>();
+        }
+
+        /// <summary>
+        /// Creates a <see cref="ClaimsIdentity"/> from the Saml2 token.
+        /// </summary>
+        /// <param name="samlToken">The Saml2SecurityToken.</param>
+        /// <param name="issuer">the issuer value for each <see cref="Claim"/> in the <see cref="ClaimsIdentity"/>.</param>
+        /// <param name="validationParameters"> contains parameters for validating the token.</param>
+        /// <returns>An IClaimIdentity.</returns>
+        protected virtual ClaimsIdentity CreateClaimsIdentity(Saml2SecurityToken samlToken, string issuer, TokenValidationParameters validationParameters)
         {
             if (samlToken == null)
             {
                 throw new ArgumentNullException("samlToken");
+            }
+
+            if (validationParameters == null)
+            {
+                throw new ArgumentNullException("validationParameters");
             }
 
             if (string.IsNullOrWhiteSpace(issuer))
@@ -121,13 +164,15 @@ namespace Microsoft.IdentityModel.Tokens
                 throw new ArgumentException(ErrorMessages.IDX10221);
             }
 
-            if (samlToken.Assertion == null)
+            Saml2Assertion assertion = samlToken.Assertion;
+            if (assertion == null)
             {
                 throw new ArgumentException(ErrorMessages.IDX10202);
             }
 
             ClaimsIdentity identity = validationParameters.CreateClaimsIdentity(samlToken, issuer);
-            _smSamlHandlerPrivateNeverSetAnyProperties.ProcessStatmentPublic(samlToken.Assertion.Statements, identity, issuer);
+            _smSaml2HandlerPrivateNeverSetAnyProperties.ProcessSamlSubjectPublic(samlToken.Assertion.Subject, identity, issuer);
+            _smSaml2HandlerPrivateNeverSetAnyProperties.ProcessStatmentPublic(samlToken.Assertion.Statements, identity, issuer);
             return identity;
         }
 
@@ -144,33 +189,7 @@ namespace Microsoft.IdentityModel.Tokens
                 throw new ArgumentNullException("tokenDescriptor");
             }
 
-            return _smSamlHandlerPrivateNeverSetAnyProperties.CreateToken(tokenDescriptor);
-        }
-
-        /// <summary>
-        /// Creates the security securityToken reference when the securityToken is not attached to the message.
-        /// </summary>
-        /// <param name="token">The saml securityToken.</param>
-        /// <param name="attached">Boolean that indicates if a attached or unattached
-        /// reference needs to be created.</param>
-        /// <returns>A <see cref="SamlAssertionKeyIdentifierClause"/>.</returns>
-        public override SecurityKeyIdentifierClause CreateSecurityTokenReference(SecurityToken token, bool attached)
-        {
-            if (null == token)
-            {
-                throw new ArgumentNullException("token");
-            }
-
-            return token.CreateKeyIdentifierClause<SamlAssertionKeyIdentifierClause>();
-        }
-
-        /// <summary>
-        /// Gets the token type identifier(s) supported by this handler.
-        /// </summary>
-        /// <returns>A collection of strings that identify the tokens this instance can handle.</returns>
-        public override string[] GetTokenTypeIdentifiers()
-        {
-            return _tokenTypeIdentifiers;
+            return _smSaml2HandlerPrivateNeverSetAnyProperties.CreateToken(tokenDescriptor);
         }
 
         /// <summary>
@@ -180,14 +199,6 @@ namespace Microsoft.IdentityModel.Tokens
         public override void LoadCustomConfiguration(XmlNodeList nodelist)
         {
             throw new NotSupportedException(ErrorMessages.IDX11004);
-        }
-
-        /// <summary>
-        /// Gets the securityToken type supported by this handler.
-        /// </summary>
-        public override Type TokenType
-        {
-            get { return typeof(SamlSecurityToken); }
         }
 
         /// <summary>
@@ -213,35 +224,46 @@ namespace Microsoft.IdentityModel.Tokens
         }
 
         /// <summary>
-        /// Obsolete method, use <see cref="ReadToken(XmlReader, TokenValidationParameters)"/> to read a <see cref="SamlSecurityToken"/>.
+        /// Obsolete method, use <see cref="ReadToken(XmlReader, TokenValidationParameters)"/> to read a <see cref="Saml2SecurityToken"/>.
         /// </summary>
         /// <param name="tokenString">not supported.</param>
-        /// <exception cref="NotSupportedException"> use use <see cref="ReadToken(XmlReader, TokenValidationParameters)"/> to read a <see cref="SamlSecurityToken"/>.</exception>
+        /// <exception cref="NotSupportedException"> use use <see cref="ReadToken(XmlReader, TokenValidationParameters)"/> to read a <see cref="Saml2SecurityToken"/>.</exception>
         public override SecurityToken ReadToken(string tokenString)
         {
-            throw new NotSupportedException(ErrorMessages.IDX11007);
+            throw new NotSupportedException(ErrorMessages.IDX11006);
         }
 
         /// <summary>
-        /// Obsolete method, use <see cref="ReadToken(XmlReader, TokenValidationParameters)"/> to read a <see cref="SamlSecurityToken"/>.
+        /// Obsolete method, use <see cref="ReadToken(XmlReader, TokenValidationParameters)"/> to read a <see cref="Saml2SecurityToken"/>.
         /// </summary>
-        /// <param name="reader">no supported.</param>
-        /// <exception cref="NotSupportedException"> use use <see cref="ReadToken(XmlReader, TokenValidationParameters)"/> to read a <see cref="SamlSecurityToken"/>.</exception>
+        /// <param name="reader">not supported.</param>
+        /// <exception cref="NotSupportedException"> use use <see cref="ReadToken(XmlReader, TokenValidationParameters)"/> to read a <see cref="Saml2SecurityToken"/>.</exception>
         public override SecurityToken ReadToken(XmlReader reader)
         {
-            throw new NotSupportedException(ErrorMessages.IDX11003);
+            throw new NotSupportedException(ErrorMessages.IDX11002);
         }
 
-
         /// <summary>
-        /// Reads a SAML 11 securityToken from the XmlReader.
+        /// Reads a SAML 2.0 token from the XmlReader.
         /// </summary>
-        /// <param name="reader">A <see cref="XmlReader"/> reader positioned at a <see cref="SamlSecurityToken"/> element.</param>
-        /// <param name="validationParameters">Contains data and information needed for reading the securityToken.</param>
-        /// <returns>An instance of a <see cref="SamlSecurityToken"/>.</returns>
+        /// <param name="reader">A <see cref="XmlReader"/> reader positioned at a <see cref="Saml2SecurityToken"/> element.</param>
+        /// <param name="validationParameters">Contains data and information needed for reading the token.</param>
+        /// <exception cref="ArgumentNullException">if 'reader' is null.</exception>
+        /// <exception cref="ArgumentNullException">if 'validationParameters' is null.</exception>
+        /// <returns>An instance of a <see cref="Saml2SecurityToken"/>.</returns>
         public virtual SecurityToken ReadToken(XmlReader reader, TokenValidationParameters validationParameters)
         {
-            return (new SamlHandler()
+            if (reader == null)
+            {
+                throw new ArgumentNullException("reader");
+            }
+
+            if (validationParameters == null)
+            {
+                throw new ArgumentNullException("validationParameters");
+            }
+
+            return (new Saml2Handler()
             {
                 Configuration = new SecurityTokenHandlerConfiguration
                 {
@@ -257,21 +279,22 @@ namespace Microsoft.IdentityModel.Tokens
         /// <exception cref="NotSupportedException"> use <see cref="ValidateToken(String, TokenValidationParameters, out SecurityToken)"/>.</exception>
         public override ReadOnlyCollection<ClaimsIdentity> ValidateToken(SecurityToken token)
         {
-            throw new NotSupportedException(ErrorMessages.IDX11001);
+            throw new NotSupportedException(ErrorMessages.IDX11000);
         }
 
         /// <summary>
-        /// Reads and validates a well formed <see cref="SamlSecurityToken"/>.
+        /// Reads and validates a well fromed Saml2 token.
         /// </summary>
-        /// <param name="securityToken">A string containing a well formed securityToken.</param>
+        /// <param name="securityToken">A Saml2 token.</param>
         /// <param name="validationParameters">Contains data and information needed for validation.</param>
-        /// <param name="validatedToken">The <see cref="Saml2SecurityToken"/> that was validated.</param>
+        /// <param name="validatedToken">The <see cref="SamlSecurityToken"/> that was validated.</param>
         /// <exception cref="ArgumentNullException">'securityToken' is null or whitespace.</exception>
         /// <exception cref="ArgumentNullException">'validationParameters' is null.</exception>
-        /// <exception cref="SecurityTokenException">'securityToken.Length' > <see cref="MaximumTokenSizeInBytes"/>.</exception>
-        /// <returns>A <see cref="ClaimsPrincipal"/> generated from the claims in the Saml securityToken.</returns>
+        /// <exception cref="ArgumentException">'securityToken.Length' > <see cref="MaximumTokenSizeInBytes"/>.</exception>
+        /// <returns>A <see cref="ClaimsPrincipal"/> generated from the claims in the Saml2 securityToken.</returns>
         public virtual ClaimsPrincipal ValidateToken(string securityToken, TokenValidationParameters validationParameters, out SecurityToken validatedToken)
         {
+            validatedToken = null;
             if (string.IsNullOrWhiteSpace(securityToken))
             {
                 throw new ArgumentNullException("securityToken");
@@ -287,30 +310,31 @@ namespace Microsoft.IdentityModel.Tokens
                 throw new ArgumentException(string.Format(CultureInfo.InvariantCulture, ErrorMessages.IDX10209, securityToken.Length, MaximumTokenSizeInBytes));
             }
 
-            // Calling System.IdentityModel.Tokens.SamlSecurityTokenHandler requires Configuration and IssuerTokenResolver be set.
-            Configuration = new SecurityTokenHandlerConfiguration
-            {
-                IssuerTokenResolver = new SecurityKeyResolver(securityToken, validationParameters),
-                MaxClockSkew = validationParameters.ClockSkew,
-            };
-
-            SamlSecurityToken samlToken;
+            Saml2SecurityToken samlToken;
             using (StringReader sr = new StringReader(securityToken))
             {
                 using (XmlDictionaryReader reader = XmlDictionaryReader.CreateDictionaryReader(XmlReader.Create(sr)))
                 {
-                    samlToken = ReadToken(reader, validationParameters) as SamlSecurityToken;
+                    samlToken = (new Saml2Handler()
+                    {
+                        Configuration = new SecurityTokenHandlerConfiguration
+                        {
+                            IssuerTokenResolver = new SecurityKeyResolver(securityToken, validationParameters),
+                            MaxClockSkew = validationParameters.ClockSkew,
+                            ServiceTokenResolver = SecurityTokenResolver.CreateDefaultSecurityTokenResolver(validationParameters.ClientDecryptionTokens, true),
+                        }
+                    }).ReadToken(reader) as Saml2SecurityToken;
                 }
+            }
+
+            if (samlToken.IssuerToken == null && validationParameters.RequireSignedTokens)
+            {
+                throw new SecurityTokenValidationException(ErrorMessages.IDX10213);
             }
 
             if (samlToken.Assertion == null)
             {
                 throw new ArgumentException(ErrorMessages.IDX10202);
-            }
-
-            if (samlToken.Assertion.SigningToken == null && validationParameters.RequireSignedTokens)
-            {
-                throw new SecurityTokenValidationException(ErrorMessages.IDX10213);
             }
 
             DateTime? notBefore = null;
@@ -341,22 +365,27 @@ namespace Microsoft.IdentityModel.Tokens
             if (validationParameters.ValidateAudience)
             {
                 List<string> audiences = new List<string>();
-                if (samlToken.Assertion.Conditions != null && samlToken.Assertion.Conditions.Conditions != null)
+                if (samlToken.Assertion.Conditions != null && samlToken.Assertion.Conditions.AudienceRestrictions != null)
                 {
-                    foreach (SamlCondition condition in samlToken.Assertion.Conditions.Conditions)
+                    foreach (Saml2AudienceRestriction restriction in samlToken.Assertion.Conditions.AudienceRestrictions)
                     {
-                        SamlAudienceRestrictionCondition audienceRestriction = condition as SamlAudienceRestrictionCondition;
-                        if (null == audienceRestriction)
+                        if (restriction == null)
                         {
                             continue;
                         }
 
-                        foreach (Uri uri in audienceRestriction.Audiences)
+                        foreach (Uri uri in restriction.Audiences)
                         {
+                            if (uri == null)
+                            {
+                                continue;
+                            }
+
                             audiences.Add(uri.OriginalString);
                         }
                     }
                 }
+
 
                 if (validationParameters.AudienceValidator != null)
                 {
@@ -371,9 +400,7 @@ namespace Microsoft.IdentityModel.Tokens
                 }
             }
 
-            string issuer = null;
-            issuer = samlToken.Assertion.Issuer == null ? null : samlToken.Assertion.Issuer;
-
+            string issuer  = samlToken.Assertion.Issuer != null ? samlToken.Assertion.Issuer.Value : null;
             if (validationParameters.ValidateIssuer)
             {
                 if (validationParameters.IssuerValidator != null)
@@ -386,9 +413,9 @@ namespace Microsoft.IdentityModel.Tokens
                 }
             }
 
-            if (samlToken.Assertion.SigningToken != null)
+            if (samlToken.IssuerToken != null)
             {
-                ValidateIssuerSecurityKey(samlToken.Assertion.SigningToken.SecurityKeys[0], samlToken, validationParameters);
+                ValidateIssuerSecurityKey(samlToken.IssuerToken.SecurityKeys[0], samlToken, validationParameters);
             }
 
             ClaimsIdentity identity = CreateClaimsIdentity(samlToken, issuer, validationParameters);
@@ -402,10 +429,10 @@ namespace Microsoft.IdentityModel.Tokens
         }
 
         /// <summary>
-        /// Determines if the audiences found in a <see cref="SamlSecurityToken"/> are valid.
+        /// Determines if the audiences found in a <see cref="Saml2SecurityToken"/> are valid.
         /// </summary>
-        /// <param name="audiences">The audiences found in the <see cref="SamlSecurityToken"/>.</param>
-        /// <param name="securityToken">The <see cref="SamlSecurityToken"/> being validated.</param>
+        /// <param name="audiences">The audiences found in the <see cref="Saml2SecurityToken"/>.</param>
+        /// <param name="securityToken">The <see cref="Saml2SecurityToken"/> being validated.</param>
         /// <param name="validationParameters"><see cref="TokenValidationParameters"/> required for validation.</param>
         /// <remarks>see <see cref="Validators.ValidateAudience"/> for additional details.</remarks>
         protected virtual void ValidateAudience(IEnumerable<string> audiences, SecurityToken securityToken, TokenValidationParameters validationParameters)
@@ -414,29 +441,29 @@ namespace Microsoft.IdentityModel.Tokens
         }
 
         /// <summary>
-        /// Determines if an issuer found in a <see cref="SamlSecurityToken"/> is valid.
+        /// Validates the lifetime of a <see cref="Saml2SecurityToken"/>.
+        /// </summary>
+        /// <param name="notBefore">The 'notBefore' time found in the <see cref="Saml2SecurityToken"/>.</param>
+        /// <param name="expires">The 'expiration' time found in the <see cref="Saml2SecurityToken"/>.</param>
+        /// <param name="securityToken">The <see cref="Saml2SecurityToken"/> being validated.</param>
+        /// <param name="validationParameters"><see cref="TokenValidationParameters"/> required for validation.</param>
+        /// <remarks><see cref="Validators.ValidateLifetime"/> for additional details.</remarks>
+        protected virtual void ValidateLifetime(DateTime? notBefore, DateTime? expires, SecurityToken securityToken, TokenValidationParameters validationParameters)
+        {
+            Validators.ValidateLifetime(notBefore: notBefore, expires: expires, securityToken: securityToken, validationParameters: validationParameters);
+        }
+
+        /// <summary>
+        /// Determines if an issuer found in a <see cref="Saml2SecurityToken"/> is valid.
         /// </summary>
         /// <param name="issuer">The issuer to validate</param>
-        /// <param name="securityToken">The <see cref="SamlSecurityToken"/> that is being validated.</param>
+        /// <param name="securityToken">The <see cref="Saml2SecurityToken"/> that is being validated.</param>
         /// <param name="validationParameters"><see cref="TokenValidationParameters"/> required for validation.</param>
         /// <returns>The issuer to use when creating the <see cref="Claim"/>(s) in the <see cref="ClaimsIdentity"/>.</returns>
         /// <remarks><see cref="Validators.ValidateIssuer"/> for additional details.</remarks>
         protected virtual string ValidateIssuer(string issuer, SecurityToken securityToken, TokenValidationParameters validationParameters)
         {
             return Validators.ValidateIssuer(issuer, securityToken, validationParameters);
-        }
-
-        /// <summary>
-        /// Validates the <see cref="SecurityToken"/> was signed by a valid <see cref="SecurityKey"/>.
-        /// </summary>
-        /// <param name="notBefore">The 'notBefore' time found in the <see cref="SamlSecurityToken"/>.</param>
-        /// <param name="expires">The 'expiration' time found in the <see cref="SamlSecurityToken"/>.</param>
-        /// <param name="securityToken">The <see cref="SamlSecurityToken"/> being validated.</param>
-        /// <param name="validationParameters"><see cref="TokenValidationParameters"/> required for validation.</param>
-        /// <remarks><see cref="Validators.ValidateLifetime"/> for additional details.</remarks>
-        protected virtual void ValidateLifetime(DateTime? notBefore, DateTime? expires, SecurityToken securityToken, TokenValidationParameters validationParameters)
-        {
-            Validators.ValidateLifetime(notBefore: notBefore, expires: expires, securityToken: securityToken, validationParameters: validationParameters);
         }
 
         /// <summary>
@@ -451,9 +478,9 @@ namespace Microsoft.IdentityModel.Tokens
         }
 
         /// <summary>
-        /// Serializes to <see cref="SamlSecurityToken"/> to a string.
+        /// Serializes to <see cref="Saml2SecurityToken"/> to a string.
         /// </summary>
-        /// <param name="token">A <see cref="SamlSecurityToken"/>.</param>
+        /// <param name="token">A <see cref="Saml2SecurityToken"/>.</param>
         public override string WriteToken(SecurityToken token)
         {
             if (token == null)
@@ -461,26 +488,26 @@ namespace Microsoft.IdentityModel.Tokens
                 throw new ArgumentNullException("token");
             }
 
-            SamlSecurityToken samlSecurityToken = token as SamlSecurityToken;
+            Saml2SecurityToken samlSecurityToken = token as Saml2SecurityToken;
             if (samlSecurityToken == null)
             {
-                throw new ArgumentException(string.Format(CultureInfo.InvariantCulture, ErrorMessages.IDX10400, this.GetType(), typeof(SamlSecurityToken), token.GetType()));
+                throw new ArgumentException(string.Format(CultureInfo.InvariantCulture, ErrorMessages.IDX10400, this.GetType(), typeof(Saml2SecurityToken), token.GetType()));
             }
 
 
             StringBuilder stringBuilder = new StringBuilder();
             using (XmlWriter xmlWriter = XmlWriter.Create(stringBuilder))
             {
-                _smSamlHandlerPrivateNeverSetAnyProperties.WriteToken(xmlWriter, token);
+                _smSaml2HandlerPrivateNeverSetAnyProperties.WriteToken(xmlWriter, token);
                 return stringBuilder.ToString();
             }
         }
 
         /// <summary>
-        /// Serializes to XML a securityToken of the type handled by this instance.
+        /// Serializes to XML a token of the type handled by this instance.
         /// </summary>
         /// <param name="writer">The XML writer.</param>
-        /// <param name="token">A securityToken of type <see cref="TokenType"/>.</param>
+        /// <param name="token">A token of type <see cref="TokenType"/>.</param>
         public override void WriteToken(XmlWriter writer, SecurityToken token)
         {
             if (writer == null)
@@ -493,20 +520,25 @@ namespace Microsoft.IdentityModel.Tokens
                 throw new ArgumentNullException("token");
             }
 
-            SamlSecurityToken samlSecurityToken = token as SamlSecurityToken;
+            Saml2SecurityToken samlSecurityToken = token as Saml2SecurityToken;
             if (samlSecurityToken == null)
             {
                 throw new ArgumentException(string.Format(CultureInfo.InvariantCulture, ErrorMessages.IDX10400, this.GetType(), typeof(SamlSecurityToken), token.GetType()));
             }
 
-            _smSamlHandlerPrivateNeverSetAnyProperties.WriteToken(writer, token);
+            _smSaml2HandlerPrivateNeverSetAnyProperties.WriteToken(writer, token);
         }
 
-        private class SMSamlHandlerPrivate : SamlHandler
+        private class SMSaml2HandlerPrivate : Saml2Handler
         {
-            public void ProcessStatmentPublic(IList<SamlStatement> statements, ClaimsIdentity subject, string issuer)
+            public void ProcessStatmentPublic(Collection<Saml2Statement> statements, ClaimsIdentity subject, string issuer)
             {
                 base.ProcessStatement(statements, subject, issuer);
+            }
+
+            public void ProcessSamlSubjectPublic(Saml2Subject assertionSubject, ClaimsIdentity subject, string issuer)
+            {
+                base.ProcessSamlSubject(assertionSubject, subject, issuer);
             }
         }
     }
