+++ src/System.IdentityModel.Tokens/Saml/SamlSecurityTokenHandler.cs
@@ -20,58 +20,40 @@ using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.Globalization;
-using System.IdentityModel.Selectors;
 using System.IdentityModel.Tokens;
 using System.IO;
 using System.Security.Claims;
 using System.Text;
 using System.Xml;
-using Saml2Handler = System.IdentityModel.Tokens.Saml2SecurityTokenHandler;
+using SamlHandler = System.IdentityModel.Tokens.SamlSecurityTokenHandler;
 
 namespace Microsoft.IdentityModel.Tokens
 {
     /// <summary>
-    /// A derived <see cref="System.IdentityModel.Tokens.Saml2SecurityTokenHandler"/> that implements ISecurityTokenValidator, 
+    /// A derived <see cref="System.IdentityModel.Tokens.Saml2SecurityTokenHandler"/> that implements ISecurityTokenValidator,
     /// which supports validating tokens passed as strings using <see cref="TokenValidationParameters"/>.
     /// </summary>
-    ///     
-    public class Saml2SecurityTokenHandler : SecurityTokenHandler, ISecurityTokenValidator
-    {        
-        internal const string Saml2TokenProfile11 = "urn:oasis:names:tc:SAML:2.0:assertion";
-        internal const string OasisWssSaml2TokenProfile11 = "http://docs.oasis-open.org/wss/oasis-wss-saml-token-profile-1.1#SAMLV2.0";
+    ///
+    public class SamlSecurityTokenHandler : SecurityTokenHandler, ISecurityTokenValidator
+    {
+        internal const string SamlTokenProfile11 = "urn:oasis:names:tc:SAML:1.0:assertion";
+        internal const string OasisWssSamlTokenProfile11 = "http://docs.oasis-open.org/wss/oasis-wss-saml-token-profile-1.1#SAMLV1.1";
 
         private Int32 _maximumTokenSizeInBytes = TokenValidationParameters.DefaultMaximumTokenSizeInBytes;
-        private static string[] _tokenTypeIdentifiers = new string[] { Saml2TokenProfile11, OasisWssSaml2TokenProfile11 };
+        private static string[] _tokenTypeIdentifiers = new string[] { SamlTokenProfile11, OasisWssSamlTokenProfile11 };
 
         // never set any properties on the handler.
-        private static SMSaml2HandlerPrivate _smSaml2HandlerPrivateNeverSetAnyProperties = new SMSaml2HandlerPrivate();
+        private static SMSamlHandlerPrivate _smSamlHandlerPrivateNeverSetAnyProperties = new SMSamlHandlerPrivate();
 
         /// <summary>
-        /// Initializes a new instance of <see cref="Saml2SecurityTokenHandler"/>.
+        /// Initializes an instance of <see cref="SamlSecurityTokenHandler"/>
         /// </summary>
-        public Saml2SecurityTokenHandler()
+        public SamlSecurityTokenHandler()
         {
         }
 
         /// <summary>
-        /// Gets the token type identifier(s) supported by this handler.
-        /// </summary>
-        /// <returns>A collection of strings that identify the tokens this instance can handle.</returns>
-        public override string[] GetTokenTypeIdentifiers()
-        {
-            return _tokenTypeIdentifiers;
-        }
-
-        /// <summary>
-        /// Gets the token type supported by this handler.
-        /// </summary>
-        public override Type TokenType
-        {
-            get { return typeof(Saml2SecurityToken); }
-        }
-
-        /// <summary>
-        /// Gets a value indicating whether this handler supports validation of tokens 
+        /// Gets a value indicating whether this handler supports validation of tokens
         /// handled by this instance.
         /// </summary>v
         /// <returns>'True' if the instance is capable of SecurityToken
@@ -82,21 +64,20 @@ namespace Microsoft.IdentityModel.Tokens
         }
 
         /// <summary>
-        /// Gets a value indicating whether the class provides serialization functionality to serialize the token handled
+        /// Gets a value indicating whether the class provides serialization functionality to serialize securityToken handled
         /// by this instance.
         /// </summary>
-        /// <returns>true if the WriteToken method can serialize this token.</returns>
+        /// <returns>true if the WriteToken method can serialize this securityToken.</returns>
         public override bool CanWriteToken
         {
             get { return true; }
         }
 
         /// <summary>
-        /// Reads the string as XML and looks for the an element <see cref="SamlConstants.Assertion"/> or  <see cref="SamlConstants.EncryptedAssertion"/> with namespace <see cref="SamlConstants.Saml2Namespace"/>. 
+        /// Reads the string as XML and looks for the an element <see cref="SamlConstants.Assertion"/> with namespace <see cref="SamlConstants.Saml11Namespace"/>.
         /// </summary>
         /// <param name="securityToken">The securitytoken.</param>
-        /// <returns><see cref="XmlDictionaryReader.IsStartElement"/> (<see cref="SamlConstants.Assertion"/>, <see cref="SamlConstants.Saml2Namespace"/>)
-        /// OR <see cref="XmlDictionaryReader.IsStartElement"/> (<see cref="SamlConstants.EncryptedAssertion"/>, <see cref="SamlConstants.Saml2Namespace"/>).</returns>
+        /// <returns><see cref="XmlDictionaryReader.IsStartElement"/> (<see cref="SamlConstants.Assertion"/>, <see cref="SamlConstants.Saml11Namespace"/>).</returns>
         public override bool CanReadToken(string securityToken)
         {
             if (string.IsNullOrWhiteSpace(securityToken) || securityToken.Length > MaximumTokenSizeInBytes)
@@ -116,47 +97,23 @@ namespace Microsoft.IdentityModel.Tokens
                     {
                         return false;
                     }
-
-                    return reader.IsStartElement(SamlConstants.Assertion, SamlConstants.Saml2Namespace)
-                        || reader.IsStartElement(SamlConstants.EncryptedAssertion, SamlConstants.Saml2Namespace);
+                    return reader.IsStartElement(SamlConstants.Assertion, SamlConstants.Saml11Namespace);
                 }
             }
         }
 
         /// <summary>
-        /// Creates the security token reference when the token is not attached to the message.
+        /// Creates claims from a Saml securityToken.
         /// </summary>
-        /// <param name="token">The saml token.</param>
-        /// <param name="attached">Boolean that indicates if a attached or unattached
-        /// reference needs to be created.</param>
-        /// <returns>A <see cref="Saml2AssertionKeyIdentifierClause"/>.</returns>
-        public override SecurityKeyIdentifierClause CreateSecurityTokenReference(SecurityToken token, bool attached)
-        {
-            if (null == token)
-            {
-                throw new ArgumentNullException("token");
-            }
-
-            return token.CreateKeyIdentifierClause<Saml2AssertionKeyIdentifierClause>();
-        }
-
-        /// <summary>
-        /// Creates a <see cref="ClaimsIdentity"/> from the Saml2 token.
-        /// </summary>
-        /// <param name="samlToken">The Saml2SecurityToken.</param>
+        /// <param name="samlToken">A <see cref="SamlSecurityToken"/> that will be used to create the claims.</param>
         /// <param name="issuer">the issuer value for each <see cref="Claim"/> in the <see cref="ClaimsIdentity"/>.</param>
-        /// <param name="validationParameters"> contains parameters for validating the token.</param>
-        /// <returns>An IClaimIdentity.</returns>
-        protected virtual ClaimsIdentity CreateClaimsIdentity(Saml2SecurityToken samlToken, string issuer, TokenValidationParameters validationParameters)
+        /// <param name="validationParameters"> contains parameters for validating the securityToken.</param>
+        /// <returns>A <see cref="ClaimsIdentity"/> containing the claims from the <see cref="SamlSecurityToken"/>.</returns>
+        protected virtual ClaimsIdentity CreateClaimsIdentity(SamlSecurityToken samlToken, string issuer, TokenValidationParameters validationParameters)
         {
             if (samlToken == null)
             {
                 throw new ArgumentNullException("samlToken");
-            }
-
-            if (validationParameters == null)
-            {
-                throw new ArgumentNullException("validationParameters");
             }
 
             if (string.IsNullOrWhiteSpace(issuer))
@@ -164,15 +121,13 @@ namespace Microsoft.IdentityModel.Tokens
                 throw new ArgumentException(ErrorMessages.IDX10221);
             }
 
-            Saml2Assertion assertion = samlToken.Assertion;
-            if (assertion == null)
+            if (samlToken.Assertion == null)
             {
                 throw new ArgumentException(ErrorMessages.IDX10202);
             }
 
             ClaimsIdentity identity = validationParameters.CreateClaimsIdentity(samlToken, issuer);
-            _smSaml2HandlerPrivateNeverSetAnyProperties.ProcessSamlSubjectPublic(samlToken.Assertion.Subject, identity, issuer);
-            _smSaml2HandlerPrivateNeverSetAnyProperties.ProcessStatmentPublic(samlToken.Assertion.Statements, identity, issuer);
+            _smSamlHandlerPrivateNeverSetAnyProperties.ProcessStatmentPublic(samlToken.Assertion.Statements, identity, issuer);
             return identity;
         }
 
@@ -189,7 +144,33 @@ namespace Microsoft.IdentityModel.Tokens
                 throw new ArgumentNullException("tokenDescriptor");
             }
 
-            return _smSaml2HandlerPrivateNeverSetAnyProperties.CreateToken(tokenDescriptor);
+            return _smSamlHandlerPrivateNeverSetAnyProperties.CreateToken(tokenDescriptor);
+        }
+
+        /// <summary>
+        /// Creates the security securityToken reference when the securityToken is not attached to the message.
+        /// </summary>
+        /// <param name="token">The saml securityToken.</param>
+        /// <param name="attached">Boolean that indicates if a attached or unattached
+        /// reference needs to be created.</param>
+        /// <returns>A <see cref="SamlAssertionKeyIdentifierClause"/>.</returns>
+        public override SecurityKeyIdentifierClause CreateSecurityTokenReference(SecurityToken token, bool attached)
+        {
+            if (null == token)
+            {
+                throw new ArgumentNullException("token");
+            }
+
+            return token.CreateKeyIdentifierClause<SamlAssertionKeyIdentifierClause>();
+        }
+
+        /// <summary>
+        /// Gets the token type identifier(s) supported by this handler.
+        /// </summary>
+        /// <returns>A collection of strings that identify the tokens this instance can handle.</returns>
+        public override string[] GetTokenTypeIdentifiers()
+        {
+            return _tokenTypeIdentifiers;
         }
 
         /// <summary>
@@ -199,6 +180,14 @@ namespace Microsoft.IdentityModel.Tokens
         public override void LoadCustomConfiguration(XmlNodeList nodelist)
         {
             throw new NotSupportedException(ErrorMessages.IDX11004);
+        }
+
+        /// <summary>
+        /// Gets the securityToken type supported by this handler.
+        /// </summary>
+        public override Type TokenType
+        {
+            get { return typeof(SamlSecurityToken); }
         }
 
         /// <summary>
@@ -224,46 +213,35 @@ namespace Microsoft.IdentityModel.Tokens
         }
 
         /// <summary>
-        /// Obsolete method, use <see cref="ReadToken(XmlReader, TokenValidationParameters)"/> to read a <see cref="Saml2SecurityToken"/>.
+        /// Obsolete method, use <see cref="ReadToken(XmlReader, TokenValidationParameters)"/> to read a <see cref="SamlSecurityToken"/>.
         /// </summary>
         /// <param name="tokenString">not supported.</param>
-        /// <exception cref="NotSupportedException"> use use <see cref="ReadToken(XmlReader, TokenValidationParameters)"/> to read a <see cref="Saml2SecurityToken"/>.</exception>
+        /// <exception cref="NotSupportedException"> use use <see cref="ReadToken(XmlReader, TokenValidationParameters)"/> to read a <see cref="SamlSecurityToken"/>.</exception>
         public override SecurityToken ReadToken(string tokenString)
         {
-            throw new NotSupportedException(ErrorMessages.IDX11006);
+            throw new NotSupportedException(ErrorMessages.IDX11007);
         }
 
         /// <summary>
-        /// Obsolete method, use <see cref="ReadToken(XmlReader, TokenValidationParameters)"/> to read a <see cref="Saml2SecurityToken"/>.
+        /// Obsolete method, use <see cref="ReadToken(XmlReader, TokenValidationParameters)"/> to read a <see cref="SamlSecurityToken"/>.
         /// </summary>
-        /// <param name="reader">not supported.</param>
-        /// <exception cref="NotSupportedException"> use use <see cref="ReadToken(XmlReader, TokenValidationParameters)"/> to read a <see cref="Saml2SecurityToken"/>.</exception>
+        /// <param name="reader">no supported.</param>
+        /// <exception cref="NotSupportedException"> use use <see cref="ReadToken(XmlReader, TokenValidationParameters)"/> to read a <see cref="SamlSecurityToken"/>.</exception>
         public override SecurityToken ReadToken(XmlReader reader)
         {
-            throw new NotSupportedException(ErrorMessages.IDX11002);
+            throw new NotSupportedException(ErrorMessages.IDX11003);
         }
 
+
         /// <summary>
-        /// Reads a SAML 2.0 token from the XmlReader.
+        /// Reads a SAML 11 securityToken from the XmlReader.
         /// </summary>
-        /// <param name="reader">A <see cref="XmlReader"/> reader positioned at a <see cref="Saml2SecurityToken"/> element.</param>
-        /// <param name="validationParameters">Contains data and information needed for reading the token.</param>
-        /// <exception cref="ArgumentNullException">if 'reader' is null.</exception>
-        /// <exception cref="ArgumentNullException">if 'validationParameters' is null.</exception>
-        /// <returns>An instance of a <see cref="Saml2SecurityToken"/>.</returns>
+        /// <param name="reader">A <see cref="XmlReader"/> reader positioned at a <see cref="SamlSecurityToken"/> element.</param>
+        /// <param name="validationParameters">Contains data and information needed for reading the securityToken.</param>
+        /// <returns>An instance of a <see cref="SamlSecurityToken"/>.</returns>
         public virtual SecurityToken ReadToken(XmlReader reader, TokenValidationParameters validationParameters)
         {
-            if (reader == null)
-            {
-                throw new ArgumentNullException("reader");
-            }
-
-            if (validationParameters == null)
-            {
-                throw new ArgumentNullException("validationParameters");
-            }
-
-            return (new Saml2Handler()
+            return (new SamlHandler()
             {
                 Configuration = new SecurityTokenHandlerConfiguration
                 {
@@ -279,22 +257,21 @@ namespace Microsoft.IdentityModel.Tokens
         /// <exception cref="NotSupportedException"> use <see cref="ValidateToken(String, TokenValidationParameters, out SecurityToken)"/>.</exception>
         public override ReadOnlyCollection<ClaimsIdentity> ValidateToken(SecurityToken token)
         {
-            throw new NotSupportedException(ErrorMessages.IDX11000);
+            throw new NotSupportedException(ErrorMessages.IDX11001);
         }
 
         /// <summary>
-        /// Reads and validates a well fromed Saml2 token.
+        /// Reads and validates a well formed <see cref="SamlSecurityToken"/>.
         /// </summary>
-        /// <param name="securityToken">A Saml2 token.</param>
+        /// <param name="securityToken">A string containing a well formed securityToken.</param>
         /// <param name="validationParameters">Contains data and information needed for validation.</param>
-        /// <param name="validatedToken">The <see cref="SamlSecurityToken"/> that was validated.</param>
+        /// <param name="validatedToken">The <see cref="Saml2SecurityToken"/> that was validated.</param>
         /// <exception cref="ArgumentNullException">'securityToken' is null or whitespace.</exception>
         /// <exception cref="ArgumentNullException">'validationParameters' is null.</exception>
-        /// <exception cref="ArgumentException">'securityToken.Length' > <see cref="MaximumTokenSizeInBytes"/>.</exception>
-        /// <returns>A <see cref="ClaimsPrincipal"/> generated from the claims in the Saml2 securityToken.</returns>
+        /// <exception cref="SecurityTokenException">'securityToken.Length' > <see cref="MaximumTokenSizeInBytes"/>.</exception>
+        /// <returns>A <see cref="ClaimsPrincipal"/> generated from the claims in the Saml securityToken.</returns>
         public virtual ClaimsPrincipal ValidateToken(string securityToken, TokenValidationParameters validationParameters, out SecurityToken validatedToken)
         {
-            validatedToken = null;
             if (string.IsNullOrWhiteSpace(securityToken))
             {
                 throw new ArgumentNullException("securityToken");
@@ -310,31 +287,30 @@ namespace Microsoft.IdentityModel.Tokens
                 throw new ArgumentException(string.Format(CultureInfo.InvariantCulture, ErrorMessages.IDX10209, securityToken.Length, MaximumTokenSizeInBytes));
             }
 
-            Saml2SecurityToken samlToken;
+            // Calling System.IdentityModel.Tokens.SamlSecurityTokenHandler requires Configuration and IssuerTokenResolver be set.
+            Configuration = new SecurityTokenHandlerConfiguration
+            {
+                IssuerTokenResolver = new SecurityKeyResolver(securityToken, validationParameters),
+                MaxClockSkew = validationParameters.ClockSkew,
+            };
+
+            SamlSecurityToken samlToken;
             using (StringReader sr = new StringReader(securityToken))
             {
                 using (XmlDictionaryReader reader = XmlDictionaryReader.CreateDictionaryReader(XmlReader.Create(sr)))
                 {
-                    samlToken = (new Saml2Handler()
-                    {
-                        Configuration = new SecurityTokenHandlerConfiguration
-                        {
-                            IssuerTokenResolver = new SecurityKeyResolver(securityToken, validationParameters),
-                            MaxClockSkew = validationParameters.ClockSkew,
-                            ServiceTokenResolver = SecurityTokenResolver.CreateDefaultSecurityTokenResolver(validationParameters.ClientDecryptionTokens, true),
-                        }
-                    }).ReadToken(reader) as Saml2SecurityToken;
+                    samlToken = ReadToken(reader, validationParameters) as SamlSecurityToken;
                 }
-            }
-
-            if (samlToken.IssuerToken == null && validationParameters.RequireSignedTokens)
-            {
-                throw new SecurityTokenValidationException(ErrorMessages.IDX10213);
             }
 
             if (samlToken.Assertion == null)
             {
                 throw new ArgumentException(ErrorMessages.IDX10202);
+            }
+
+            if (samlToken.Assertion.SigningToken == null && validationParameters.RequireSignedTokens)
+            {
+                throw new SecurityTokenValidationException(ErrorMessages.IDX10213);
             }
 
             DateTime? notBefore = null;
@@ -365,27 +341,22 @@ namespace Microsoft.IdentityModel.Tokens
             if (validationParameters.ValidateAudience)
             {
                 List<string> audiences = new List<string>();
-                if (samlToken.Assertion.Conditions != null && samlToken.Assertion.Conditions.AudienceRestrictions != null)
+                if (samlToken.Assertion.Conditions != null && samlToken.Assertion.Conditions.Conditions != null)
                 {
-                    foreach (Saml2AudienceRestriction restriction in samlToken.Assertion.Conditions.AudienceRestrictions)
+                    foreach (SamlCondition condition in samlToken.Assertion.Conditions.Conditions)
                     {
-                        if (restriction == null)
+                        SamlAudienceRestrictionCondition audienceRestriction = condition as SamlAudienceRestrictionCondition;
+                        if (null == audienceRestriction)
                         {
                             continue;
                         }
 
-                        foreach (Uri uri in restriction.Audiences)
+                        foreach (Uri uri in audienceRestriction.Audiences)
                         {
-                            if (uri == null)
-                            {
-                                continue;
-                            }
-
                             audiences.Add(uri.OriginalString);
                         }
                     }
                 }
-
 
                 if (validationParameters.AudienceValidator != null)
                 {
@@ -400,7 +371,9 @@ namespace Microsoft.IdentityModel.Tokens
                 }
             }
 
-            string issuer  = samlToken.Assertion.Issuer != null ? samlToken.Assertion.Issuer.Value : null;
+            string issuer = null;
+            issuer = samlToken.Assertion.Issuer == null ? null : samlToken.Assertion.Issuer;
+
             if (validationParameters.ValidateIssuer)
             {
                 if (validationParameters.IssuerValidator != null)
@@ -413,9 +386,9 @@ namespace Microsoft.IdentityModel.Tokens
                 }
             }
 
-            if (samlToken.IssuerToken != null)
+            if (samlToken.Assertion.SigningToken != null)
             {
-                ValidateIssuerSecurityKey(samlToken.IssuerToken.SecurityKeys[0], samlToken, validationParameters);
+                ValidateIssuerSecurityKey(samlToken.Assertion.SigningToken.SecurityKeys[0], samlToken, validationParameters);
             }
 
             ClaimsIdentity identity = CreateClaimsIdentity(samlToken, issuer, validationParameters);
@@ -429,10 +402,10 @@ namespace Microsoft.IdentityModel.Tokens
         }
 
         /// <summary>
-        /// Determines if the audiences found in a <see cref="Saml2SecurityToken"/> are valid.
+        /// Determines if the audiences found in a <see cref="SamlSecurityToken"/> are valid.
         /// </summary>
-        /// <param name="audiences">The audiences found in the <see cref="Saml2SecurityToken"/>.</param>
-        /// <param name="securityToken">The <see cref="Saml2SecurityToken"/> being validated.</param>
+        /// <param name="audiences">The audiences found in the <see cref="SamlSecurityToken"/>.</param>
+        /// <param name="securityToken">The <see cref="SamlSecurityToken"/> being validated.</param>
         /// <param name="validationParameters"><see cref="TokenValidationParameters"/> required for validation.</param>
         /// <remarks>see <see cref="Validators.ValidateAudience"/> for additional details.</remarks>
         protected virtual void ValidateAudience(IEnumerable<string> audiences, SecurityToken securityToken, TokenValidationParameters validationParameters)
@@ -441,29 +414,29 @@ namespace Microsoft.IdentityModel.Tokens
         }
 
         /// <summary>
-        /// Validates the lifetime of a <see cref="Saml2SecurityToken"/>.
-        /// </summary>
-        /// <param name="notBefore">The 'notBefore' time found in the <see cref="Saml2SecurityToken"/>.</param>
-        /// <param name="expires">The 'expiration' time found in the <see cref="Saml2SecurityToken"/>.</param>
-        /// <param name="securityToken">The <see cref="Saml2SecurityToken"/> being validated.</param>
-        /// <param name="validationParameters"><see cref="TokenValidationParameters"/> required for validation.</param>
-        /// <remarks><see cref="Validators.ValidateLifetime"/> for additional details.</remarks>
-        protected virtual void ValidateLifetime(DateTime? notBefore, DateTime? expires, SecurityToken securityToken, TokenValidationParameters validationParameters)
-        {
-            Validators.ValidateLifetime(notBefore: notBefore, expires: expires, securityToken: securityToken, validationParameters: validationParameters);
-        }
-
-        /// <summary>
-        /// Determines if an issuer found in a <see cref="Saml2SecurityToken"/> is valid.
+        /// Determines if an issuer found in a <see cref="SamlSecurityToken"/> is valid.
         /// </summary>
         /// <param name="issuer">The issuer to validate</param>
-        /// <param name="securityToken">The <see cref="Saml2SecurityToken"/> that is being validated.</param>
+        /// <param name="securityToken">The <see cref="SamlSecurityToken"/> that is being validated.</param>
         /// <param name="validationParameters"><see cref="TokenValidationParameters"/> required for validation.</param>
         /// <returns>The issuer to use when creating the <see cref="Claim"/>(s) in the <see cref="ClaimsIdentity"/>.</returns>
         /// <remarks><see cref="Validators.ValidateIssuer"/> for additional details.</remarks>
         protected virtual string ValidateIssuer(string issuer, SecurityToken securityToken, TokenValidationParameters validationParameters)
         {
             return Validators.ValidateIssuer(issuer, securityToken, validationParameters);
+        }
+
+        /// <summary>
+        /// Validates the <see cref="SecurityToken"/> was signed by a valid <see cref="SecurityKey"/>.
+        /// </summary>
+        /// <param name="notBefore">The 'notBefore' time found in the <see cref="SamlSecurityToken"/>.</param>
+        /// <param name="expires">The 'expiration' time found in the <see cref="SamlSecurityToken"/>.</param>
+        /// <param name="securityToken">The <see cref="SamlSecurityToken"/> being validated.</param>
+        /// <param name="validationParameters"><see cref="TokenValidationParameters"/> required for validation.</param>
+        /// <remarks><see cref="Validators.ValidateLifetime"/> for additional details.</remarks>
+        protected virtual void ValidateLifetime(DateTime? notBefore, DateTime? expires, SecurityToken securityToken, TokenValidationParameters validationParameters)
+        {
+            Validators.ValidateLifetime(notBefore: notBefore, expires: expires, securityToken: securityToken, validationParameters: validationParameters);
         }
 
         /// <summary>
@@ -478,9 +451,9 @@ namespace Microsoft.IdentityModel.Tokens
         }
 
         /// <summary>
-        /// Serializes to <see cref="Saml2SecurityToken"/> to a string.
+        /// Serializes to <see cref="SamlSecurityToken"/> to a string.
         /// </summary>
-        /// <param name="token">A <see cref="Saml2SecurityToken"/>.</param>
+        /// <param name="token">A <see cref="SamlSecurityToken"/>.</param>
         public override string WriteToken(SecurityToken token)
         {
             if (token == null)
@@ -488,26 +461,26 @@ namespace Microsoft.IdentityModel.Tokens
                 throw new ArgumentNullException("token");
             }
 
-            Saml2SecurityToken samlSecurityToken = token as Saml2SecurityToken;
+            SamlSecurityToken samlSecurityToken = token as SamlSecurityToken;
             if (samlSecurityToken == null)
             {
-                throw new ArgumentException(string.Format(CultureInfo.InvariantCulture, ErrorMessages.IDX10400, this.GetType(), typeof(Saml2SecurityToken), token.GetType()));
+                throw new ArgumentException(string.Format(CultureInfo.InvariantCulture, ErrorMessages.IDX10400, this.GetType(), typeof(SamlSecurityToken), token.GetType()));
             }
 
 
             StringBuilder stringBuilder = new StringBuilder();
             using (XmlWriter xmlWriter = XmlWriter.Create(stringBuilder))
             {
-                _smSaml2HandlerPrivateNeverSetAnyProperties.WriteToken(xmlWriter, token);
+                _smSamlHandlerPrivateNeverSetAnyProperties.WriteToken(xmlWriter, token);
                 return stringBuilder.ToString();
             }
         }
 
         /// <summary>
-        /// Serializes to XML a token of the type handled by this instance.
+        /// Serializes to XML a securityToken of the type handled by this instance.
         /// </summary>
         /// <param name="writer">The XML writer.</param>
-        /// <param name="token">A token of type <see cref="TokenType"/>.</param>
+        /// <param name="token">A securityToken of type <see cref="TokenType"/>.</param>
         public override void WriteToken(XmlWriter writer, SecurityToken token)
         {
             if (writer == null)
@@ -520,25 +493,20 @@ namespace Microsoft.IdentityModel.Tokens
                 throw new ArgumentNullException("token");
             }
 
-            Saml2SecurityToken samlSecurityToken = token as Saml2SecurityToken;
+            SamlSecurityToken samlSecurityToken = token as SamlSecurityToken;
             if (samlSecurityToken == null)
             {
                 throw new ArgumentException(string.Format(CultureInfo.InvariantCulture, ErrorMessages.IDX10400, this.GetType(), typeof(SamlSecurityToken), token.GetType()));
             }
 
-            _smSaml2HandlerPrivateNeverSetAnyProperties.WriteToken(writer, token);
+            _smSamlHandlerPrivateNeverSetAnyProperties.WriteToken(writer, token);
         }
 
-        private class SMSaml2HandlerPrivate : Saml2Handler
+        private class SMSamlHandlerPrivate : SamlHandler
         {
-            public void ProcessStatmentPublic(Collection<Saml2Statement> statements, ClaimsIdentity subject, string issuer)
+            public void ProcessStatmentPublic(IList<SamlStatement> statements, ClaimsIdentity subject, string issuer)
             {
                 base.ProcessStatement(statements, subject, issuer);
-            }
-
-            public void ProcessSamlSubjectPublic(Saml2Subject assertionSubject, ClaimsIdentity subject, string issuer)
-            {
-                base.ProcessSamlSubject(assertionSubject, subject, issuer);
             }
         }
     }
