+++ src/System.IdentityModel.Tokens/Saml/SecurityKeyResolver.cs
