+++ src/System.IdentityModel.Tokens/Tokens/SecurityTokenHandlerCollectionExtensions.cs
